refactoring: remove static

+++ Keeper/Assets/Scripts/Avocado/Data/Components/WeaponComponentData.cs
@@ -5,13 +5,13 @@ namespace Avocado.Game.Data.Components {
     [UsedImplicitly]
     [ComponentType(ComponentType.Weapon)]
     public struct WeaponComponentData : IComponentData {
-        public readonly float Damage;
+        public readonly int Damage;
         public readonly float Delay;
         public readonly int Clip;
         public readonly float Range;
 
         public WeaponComponentData(JObject data) {
-            Damage = data["Damage"].Value<float>();
+            Damage = data["Damage"].Value<int>();
             Delay = data["Delay"].Value<float>();
             Clip = data["Clip"].Value<int>();
             Range = data["Range"].Value<float>();
